requirements.txt
streamlit
tensorflow
keras
pillow
opencv-python-headless
